Moved downloadstatus handling in modal to effects/states
