+++ src/components/widgets/DownloadButton.js
@@ -4,7 +4,7 @@ import { useSelector, useDispatch } from "react-redux";
 import { jsx } from "@emotion/core";
 import textStyles from "../../styles/text";
 import { actionCreators } from "../../redux/actionCreators";
-import DownloadModal from "../modals/downloadModal";
+import DownloadModal from "../modals/DownloadModal";
 import useModal from "../../hooks/useModal";
 import ActionButton from "./ActionButton";
